+++ tdmclient/tools/server.py
@@ -20,6 +20,7 @@ Options:
   --help         display this help message and exit
   --port=P       port (default: {Server.PORT} for TCP, {ServerWS.PORT} for WebSocket)
   --ws           WebSocket instead of plain TCP
+  --zeroconf       advertise TCP TDM port via zeroconf
 """)
 
 
@@ -27,6 +28,7 @@ if __name__ == "__main__":
 
     tdm_port = None
     ws = False
+    adv_zeroconf = False
     debug = False
 
     try:
@@ -37,6 +39,7 @@ if __name__ == "__main__":
                                               "help",
                                               "port=",
                                               "ws",
+                                              "zeroconf",
                                           ])
     except getopt.error as err:
         print(str(err))
@@ -51,6 +54,31 @@ if __name__ == "__main__":
             tdm_port = int(val)
         elif arg == "--ws":
             ws = True
+        elif arg == "--zeroconf":
+            adv_zeroconf = True
+
+    if adv_zeroconf:
+        from zeroconf import Zeroconf, ServiceInfo
+        import socket
+        import uuid
+
+        service_props = {
+            "uuid": str(uuid.uuid4())
+        }
+        if ws:
+            service_props["ws-port"] = str(tdm_port or ServerWS.PORT)
+
+        info = ServiceInfo(
+            "_mobsya._tcp.local.",
+            "Python tdmclient_tools_server._mobsya._tcp.local.",
+            weight=100,
+            addresses=[socket.inet_aton("127.0.0.1")],
+            port=tdm_port or Server.PORT,
+            properties=service_props,
+        )
+
+        zeroconf = Zeroconf()
+        zeroconf.register_service(info)
 
     node = ServerNode(type=ThymioFB.NODE_TYPE_THYMIO2,
                       variables={